'is dependent on' precedence rule
